Don't leave rev bodies behind after a merge
When the replicator inserts a revision, it marks it with the kKeepBody
flag, so the body sticks around to be used as the source of a delta
when you later push back a new revision.
When we resolve a conflict, we aren't clearing that flag on all
revisions that are not the current server revision. There are cases
where it gets left behind on an earlier rev. Fixed this.

Fixes CBL-708